Move `hull_crop` from `mesh.py` to `spatial.py`
